% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitable.R
\name{epitable_append}
\alias{epitable_append}
\title{Append multiple epitables}
\usage{
epitable_append(..., file)
}
\arguments{
\item{...}{Comma separated tables.}
}
\description{
Append multiple epitables together into a single table.
}
\keyword{html}
\keyword{tables}
