% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_by.R
\name{freq_by}
\alias{freq_by}
\title{Frequencies and percentages}
\usage{
freq_by(dataset, var_vector, by_group = NULL, include_total = TRUE,
  min_cell_count = 10, htmlout = TRUE, font_css = "font-family: Times;")
}
\arguments{
\item{dataset}{A dataset}

\item{var_vector}{A character vector containing names of the columns in the
\code{dataset} to calculate frequencies and percentages for.}

\item{by_group}{A string referring to a factor column in the \code{dataset}
by which to stratify the calculations.}

\item{include_total}{whether to include two extra columns of marginal
frequencies and percentages, i.e. not stratified by the \code{by_group}.}

\item{min_cell_count}{a number which defaults to 10. Used to preserve
anonymity in case of sensitive data. In cells with <= 10 observations,
the string "<=10" is printed.}

\item{htmlout}{Whether to output to html (default and intended usage), or
as r-dataframe.}

\item{font_css}{A string of CSS code defining the font used for the table.
Default is'font-family: monospace;'.}
}
\description{
\code{freq_by} creates frequency and percentage tables in HTML.
}
\details{
The output is a table in HTML which can be viewed in a browser or
included in a knitr-report.
}
\examples{
# Outputs HTML:
output <-
freq_by(example_data, c("cut", "color"), "clarity", htmlout = FALSE, min_cell_count = 30)
}
