% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_freq_function.R
\name{freq_function}
\alias{freq_function}
\title{Frequencies and percentages}
\usage{
freq_function(dataset, var_vektor)
}
\arguments{
\item{dataset}{A dataset}

\item{var_vector}{For internal use. A character vector containing names of the columns in the
input dataset to calculate frequencies and percentages for.}
}
\description{
\code{freq_function} creates frequency and percentage tables.
}
\keyword{##}
\keyword{-}
\keyword{AN}
\keyword{BECAUSE}
\keyword{BUILD}
\keyword{FREQ_BY}
\keyword{FUNCTION}
\keyword{IGNORE}
\keyword{INSIDE}
\keyword{IS}
\keyword{IT}
\keyword{NOW}
\keyword{SET}
\keyword{TO}
\keyword{can}
\keyword{freq_by.}
\keyword{from}
\keyword{inherit}
\keyword{internal}
\keyword{it}
\keyword{so}
