% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_becomes_html.R
\name{model_becomes_html}
\alias{model_becomes_html}
\title{Model becomes HTML}
\usage{
model_becomes_html(tidy_model, env = parent.frame())
}
\arguments{
\item{tidy_model}{A dataframe.}
}
\description{
Takes as input the output from \code{model_gets_formatted_numbers}.
}
\keyword{internal}
