% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gets_formatted_numbers.R
\name{model_gets_formatted_numbers}
\alias{model_gets_formatted_numbers}
\title{Model becomes HTML}
\usage{
model_gets_formatted_numbers(univariate_models, env = parent.frame())
}
\arguments{
\item{tidy_model}{A dataframe.}
}
\description{
Takes as input the output from \code{model_gets_formatted_numbers}.
}
\keyword{internal}
