% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gets_ref_levels.R
\name{model_gets_ref_levels}
\alias{model_gets_ref_levels}
\title{Add reference levels}
\usage{
model_gets_ref_levels(model_object, env = parent.frame())
}
\arguments{
\item{model_object}{A cox.ph object}

\item{exponentiate}{defaults to NULL. Is set to FALSE if model input is class
glm or lm, and is set to TRUE if model is class coxph.}
}
\description{
\code{add_reference_levels} Adds reference levels to regression
    model object
}
\details{
Prints summary statistics of models of class \code{coxph} or
    \code{glm} using the \code{broom}-package, and adds a reference
    level to the categorical predictors, as is commonly used in
    epidemiological publications. Does not support models with interactions
    or covariates of class \code{ordered}.
}
\examples{
# Model for which I want reference levels printed in the output.
lung <- survival::lung
lung$species <- iris$Species[ sample.int( 150, 228, replace = TRUE )]
input_to_function <- survival::coxph( survival::Surv( time, status == 2 ) ~
age + sex  + ph.karno + wt.loss + species, data =  lung)
add_reference_levels( model_object = input_to_function)
}
\keyword{internal}
