% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_html.R
\name{model_to_html}
\alias{model_to_html}
\title{Print regression models to HTML}
\usage{
model_to_html(univariate_models_list, multivariate_models_list,
  decimals_estimate = 2, exponentiate = FALSE,
  font_css = "font-family: monospace;")
}
\arguments{
\item{univariate_models_list}{List of univariate models}

\item{multivariate_models_list}{List of univariate models}

\item{decimals_estimate}{A number specifying decimals on estimates. Default is 2.}

\item{font_css}{A string of CSS code defining the font used for the table. Default is
'font-family: monospace;'. A monospace font necessary for perfect
alignment of table cells. Another option:'font-family: "Courier New";'}
}
\description{
Prints on or more lists of regression models to HTML. Currently
    only works with class coxph.
}
\examples{
df         <- survival::lung
df$age_bin <- Hmisc::cut2( df$age, g = 5)
df$ph_bin  <- Hmisc::cut2( df$ph.karno, g = 5)
df$sex     <- factor( df$sex)
model1     <- survival::coxph( survival::Surv( time , status==1) ~ ph_bin + wt.loss, data = df)

diamonds         <- example_data
diamonds$color   <- factor(diamonds$color, ordered = FALSE)
diamonds$clarity <- factor(diamonds$clarity, ordered = FALSE)
glm_logistic     <- glm( cut=="Ideal" ~  color + clarity + x , data = diamonds, family = "binomial")
glm_linear       <- glm( Sepal.Width ~  Petal.Width + Species, data = iris)
}
