% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{The titanic dataset}
\format{A data frame with 891 rows and 12 variables:
\describe{
  \item{PassengerId}{}
  \item{Survived}{}
  \item{Pclass}{}
  \item{Name}{}
  \item{Sex}{}
  \item{Age}{}
  \item{SibSp}{}
  \item{Parch}{}
  \item{Ticket}{}
  \item{Fare}{}
  \item{Cabin}{}
  \item{Embarked}{}
}}
\usage{
titanic
}
\description{
The titanic dataset from the kaggle competition, but with labels.
}
\keyword{datasets}
