% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitable.R
\name{epitable}
\alias{epitable}
\alias{knit_print.epitable}
\alias{print.epitable}
\title{A EPI HTML Table Making Function}
\usage{
epitable(x, rownamesvar, rownames, colnames = NULL, colgroups = NULL,
  colgroupspattern = NULL, select = NULL, rowlevels = NULL,
  header = TRUE, file = NULL, selfcontained = FALSE)

\method{knit_print}{epitable}(x, ...)

\method{print}{epitable}(x, useViewer = TRUE, ...)
}
\arguments{
\item{x}{is the data}

\item{file}{is the filename for saving the table snippet as a file. By default, this is NULL and epitable() does not write to a file.}

\item{selfcontained}{If writing to file, the default selfcontained=FALSE will write a table snippet. selfcontained=TRUE writes a complete html page, with header, body, etc.}

\item{useViewer}{Set to false to show snippet in console.}

\item{example}{Just for testing purposes.}
}
\description{
This is the table making function.
}
\examples{
epitable()

}
\keyword{html}
\keyword{tables}
