% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitable.R
\name{epitable}
\alias{epitable}
\alias{print.epitable}
\title{A EPI HTML Table Making Function}
\usage{
epitable(x, file = NULL, selfcontained = FALSE, example = FALSE)

\method{print}{epitable}(x, useViewer = TRUE, ...)
}
\arguments{
\item{x}{is the data}

\item{file}{is the filename for saving the table snippet as a file. By default, this is NULL and epitable() does not write to a file.}

\item{selfcontained}{If writing to file, the default selfcontained=FALSE will write a table snippet. selfcontained=TRUE writes a complete html page, with header, body, etc.}

\item{example}{Just for testing purposes.}

\item{useViewer}{If you are using RStudio there is a viewer thar can render
the table within that is envoced if in \code{\link[base]{interactive}} mode.
Set this to \code{FALSE} if you want to remove that  functionality. You can
also force the function to call a specific viewer by setting this to a
viewer function, e.g. \code{useViewer = utils::browseURL} if you want to
override the default RStudio viewer. Another option that does the same is to
set the \code{options(viewer=utils::browseURL)} and it will default to that
particular viewer (this is how RStudio decides on a viewer).
\emph{Note:} If you want to force all output to go through the
\code{\link[base]{cat}()} the set \code{\link[base]{options}(htmlTable.cat = TRUE)}.}
}
\description{
This is the table making function.
}
\examples{
epitable()

}
\keyword{html}
\keyword{tables}
