% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_freq_function.R
\name{freq_function}
\alias{freq_function}
\title{Frequencies and percentages}
\usage{
freq_function(dataset, var_vektor)
}
\arguments{
\item{dataset}{A dataset}

\item{var_vector}{For internal use. A character vector containing names of the columns in the
input dataset to calculate frequencies and percentages for.}
}
\description{
\code{freq_function} creates frequency and percentage tables.
}
\keyword{internal}
