% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_to_html.R
\name{model_to_html}
\alias{model_to_html}
\title{Print regression models to HTML}
\usage{
model_to_html(univariate_models_list, simple = TRUE, decimals_estimate = 2)
}
\arguments{
\item{univariate_models_list}{List of univariate models}

\item{decimals_estimate}{A number specifying decimals on estimates. Default is 2.}
}
\description{
Prints on or more lists of regression models to HTML. Currently
    only works with class coxph.
}
\examples{
library(survival)
df         <- lung
df$age_bin <- Hmisc::cut2( df$age, g = 5)
df$ph_bin  <- Hmisc::cut2( df$ph.karno, g = 5)
df$sex     <- factor( df$sex)
model1     <- survival::coxph( survival::Surv( time = time, event = status==1) ~ age_bin + factor(sex) + ph_bin + wt.loss, data = df )

diamonds         <- ggplot2::diamonds
diamonds$color   <- factor(diamonds$color, ordered = FALSE)
diamonds$clarity <- factor(diamonds$clarity, ordered = FALSE)
glm_logistic     <- glm( cut=="Ideal" ~  color + clarity + x , data = diamonds, family = "binomial")
glm_linear       <- glm( Sepal.Width ~  Petal.Width + Species, data = iris)
}
